+++ crates/store/src/backend/mysql/write.rs
@@ -117,7 +117,7 @@ impl MysqlStore {
                 Operation::Collection {
                     collection: collection_,
                 } => {
-                    collection = *collection_;
+                    collection = u8::from(*collection_);
                 }
                 Operation::DocumentId {
                     document_id: document_id_,
@@ -314,7 +314,7 @@ impl MysqlStore {
                 Operation::Log { collection, set } => {
                     let key = LogKey {
                         account_id,
-                        collection: *collection,
+                        collection: u8::from(*collection),
                         change_id,
                     }
                     .serialize(0);
