+++ crates/store/src/backend/postgres/write.rs
@@ -119,7 +119,7 @@ impl PostgresStore {
                 Operation::Collection {
                     collection: collection_,
                 } => {
-                    collection = *collection_;
+                    collection = u8::from(*collection_);
                 }
                 Operation::DocumentId {
                     document_id: document_id_,
@@ -315,7 +315,7 @@ impl PostgresStore {
                 Operation::Log { collection, set } => {
                     let key = LogKey {
                         account_id,
-                        collection: *collection,
+                        collection: u8::from(*collection),
                         change_id,
                     }
                     .serialize(0);
